Responsabilidade tirada dos métodos da rota
